// apps/admin/src/api/generated/model/feedbackConversationDetailDtoOutputControlMode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationDetailDtoOutputControlMode =
  (typeof FeedbackConversationDetailDtoOutputControlMode)[keyof typeof FeedbackConversationDetailDtoOutputControlMode];

export const FeedbackConversationDetailDtoOutputControlMode = {
  bot: "bot",
  human: "human",
} as const;