// apps/admin/src/api/generated/model/feedbackConversationDetailDtoOutputGoalsItemStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationDetailDtoOutputGoalsItemStatus =
  (typeof FeedbackConversationDetailDtoOutputGoalsItemStatus)[keyof typeof FeedbackConversationDetailDtoOutputGoalsItemStatus];

export const FeedbackConversationDetailDtoOutputGoalsItemStatus = {
  pending: "pending",
  asked: "asked",
  answered: "answered",
  skipped: "skipped",
} as const;