// apps/admin/src/api/generated/model/feedbackConversationDetailDtoOutputLifecycleState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationDetailDtoOutputLifecycleState =
  (typeof FeedbackConversationDetailDtoOutputLifecycleState)[keyof typeof FeedbackConversationDetailDtoOutputLifecycleState];

export const FeedbackConversationDetailDtoOutputLifecycleState = {
  open: "open",
  closed: "closed",
} as const;